Add OriginalDemonstrateApi and refactor existing code
A new file 'OriginalDemonstrateApi' has been added to demonstrate API usage. Also, redundant code was removed from 'Program.cs', and a minor change was made in 'DemonstrateAPI.cs' to simplify the initialization of '_api' field.
